+++ src/test/skript/tests/syntaxes/conditions/CondIsActivate.sk
